% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.gui.R
\name{ggplot.gui}
\alias{ggplot.gui}
\title{Creating a graphical user interface for creating ggplot-graphs.}
\usage{
\method{ggplot}{gui}(dataset = NA)
}
\arguments{
\item{dataset}{A dataset (optional).}
}
\value{
A GUI for visualizing data from \code{dataset}.
}
\description{
Creating a graphical user interface for creating ggplot-graphs.
}
\examples{
ggplot.gui()
ggplot.gui(mpg)
}
