% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_shiny.R
\name{ggplot_shiny}
\alias{ggplot_shiny}
\title{Creating a graphical user interface for creating ggplot-graphs.}
\usage{
ggplot_shiny(dataset = NA)
}
\arguments{
\item{dataset}{A dataset (optional).}
}
\value{
A GUI for visualizing data from \code{dataset}.
}
\description{
Creating a graphical user interface for creating ggplot-graphs.
}
\examples{
#ggplot_shiny()
#ggplot_shiny(mpg)
}
