% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tut_data}
\alias{tut_data}
\title{Data for ARTS102 tutorial.}
\format{
A data frame with 148 rows and 12 variables
}
\usage{
tut_data
}
\description{
Data for ARTS102 tutorial.
}
\keyword{datasets}
