% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tut_data}
\alias{tut_data}
\title{Data for ARTS102 tutorial.}
\format{
A data frame with 148 rows and 12 variables:
\describe{
  \item{Do you have any children?}{'yes' or 'no'.}
  \item{knittingbinary}{'yes' or 'no'.}
  \item{minecraftbinary}{'above' or 'below'.}
  \item{The red car is very fast}{Score for question.}
  \item{There's cats in that box}{Score for question.}
  \item{He has baked beans in his cupboard}{Score for question.}
  \item{John done that yesterday}{Score for question.}
  \item{Mary's got a red car}{Score for question.}
  \item{If you got skates, bring them along}{Score for question.}
  \item{If that person is serious about NCEA they'll have to study}{Score for question.}
  \item{The express lane is only for people with 10 items or less}{Score for question.}
  \item{If Mary wants to pass the exam, then they'll have to study.}{Score for question.}
}
}
\usage{
tut_data
}
\description{
Data for ARTS102 tutorial.
}
\keyword{datasets}
